+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Fix(es):
 
 - Add storage infos from lshw outputs to GCP servers.
 - Check region availability for Alibaba Cloud servers before adding their prices.
+- Determine CPU allocation type for Alibaba Cloud servers.
 
 ## v0.3.3 (Jan 02, 2026)